function/__init__.py
import azure.functions as func
import nest_asyncio

from src.init import app

nest_asyncio.apply()

async def main(req: func.HttpRequest, context: func.Context) -> func.HttpResponse:
    return func.AsgiMiddleware(app).handle(req, context)