requirements.txt
eth_account==0.6.1
fastapi==0.78.0
nest_asyncio==1.5.5
pydantic==1.9.1
python-dotenv==0.20.0
starlette==0.19.1
web3==6.0.0b3
azure-cosmos